// apps/desktop/src/generated/TaskManagerProcessSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskManagerProcessIdentity } from "./TaskManagerProcessIdentity";
import type { TaskManagerProcessState } from "./TaskManagerProcessState";

export type TaskManagerProcessSummary = {
  identity: TaskManagerProcessIdentity;
  parent_pid: number;
  name: string;
  state: TaskManagerProcessState;
  cpu_percent?: number | null;
  resident_memory_bytes: number;
  resident_memory_percent: number;
  virtual_memory_bytes: number;
  disk_read_bytes_per_second?: number | null;
  disk_write_bytes_per_second?: number | null;
  io_permission_denied: boolean;
  thread_count: number;
  nice: number;
};